+++ app/utils/backend/vatsim/atc-duplicating.ts
@@ -250,6 +250,23 @@ export const duplicatingSettings = [
         },
     },
     /**
+     * @description ZDV Center and TRACONs
+     * @author 1378019
+     */
+    {
+        regex: /^(DEN|D01|GJT|PUB|RCA|COS|CYS|CPR|ASE)(_\w{0,3})?_(CTR|TMU|APP|DEP)$$/,
+        mapping: {
+            DEN: 'DEN_APP',
+            GJT: 'GJT_APP',
+            PUB: 'PUB_APP',
+            RCA: 'RCA_APP',
+            COS: 'COS_APP',
+            CYS: 'CYS_APP',
+            ASE: 'ASE_APP',
+            CPR: 'CPR_APP',
+        },
+    },
+    /**
      * @description CZYZ FIR Terminal Sectors
      * @author 1401686 and 1448618
      */