+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Roadmap has been updated
 - New icons from DotWallop: AA5, AN26, B52, BE33, C46, GA7C, L159, LEG2, P06T, VL3
 - Added model matching: AN32→AN26, VL3T→VL3, PA30→PA39
+- Add CZYZ FIR Terminal Sectors to ATC Duplicating
+- Update regex for LECB area in ATC duplication
 
 # [1.2.4-6]